Move transfer recorder to its own file
